pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stereocomplex"
version = "0.1.0"
description = "Stereo calibration + compact ray-field/mapping with CPU fallback simulator."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pillow>=10.0",
  "scipy>=1.10",
]

[project.optional-dependencies]
docs = [
  "sphinx>=7.2",
  "sphinx-rtd-theme>=2.0",
  "myst-parser>=2.0",
  "sphinx-copybutton>=0.5",
  "linkify-it-py>=2.0",
  "matplotlib>=3.8",
]
dev = [
  "pytest>=7.0",
  "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
